requirements.txt
# Core dependencies
numpy>=1.21.0
opencv-python>=4.5.0
scikit-image>=0.18.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
Pillow>=8.3.0

# Jupyter
jupyter>=1.0.0
jupyterlab>=3.0.0
ipykernel>=6.0.0  # Required for running Jupyter notebooks with venv kernel

# Data processing
pandas>=1.3.0

# Deep learning dependencies
torch>=2.8.0
torchvision>=0.23.0
# tensorflow>=2.7.0  # Optional: Uncomment if using TensorFlow-based models

